+++ tests/basic.test.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable max-classes-per-file */
 
 import { act, cleanup, fireEvent, render } from '@testing-library/react';
-import { spyElementPrototypes } from 'rc-util/lib/test/domHook';
+import { spyElementPrototypes } from 'rc-util-modern/dist/test/domHook';
 import React, { StrictMode, createRef } from 'react';
 import ReactDOM from 'react-dom';
 import Trigger from '../src';