+++ tests/flipShift.test.tsx
@@ -1,6 +1,5 @@
 import { act, cleanup, render } from '@testing-library/react';
-import { spyElementPrototypes } from 'rc-util/lib/test/domHook';
-import React from 'react';
+import { spyElementPrototypes } from 'rc-util-modern/dist/test/domHook';
 import Trigger from '../src';
 
 /*