+++ tests/portal.test.jsx
@@ -1,8 +1,7 @@
 /* eslint-disable max-classes-per-file */
 
 import { act, cleanup, fireEvent, render } from '@testing-library/react';
-import { spyElementPrototypes } from 'rc-util/lib/test/domHook';
-import React from 'react';
+import { spyElementPrototypes } from 'rc-util-modern/dist/test/domHook';
 import ReactDOM from 'react-dom';
 import Trigger from '../src';
 import { placementAlignMap } from './util';